pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["sangjinye <tkdwlsdlzzz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mongoengine = "^0.20.0"
flask-classful = "^0.14.2"
marshmallow = "^3.7.1"
Flask-Cors = "^3.0.8"
flask = "1.1.1"
Pyjwt = "^1.7.1"
bcrypt = "^3.2.0"
flask_apispec = "^0.10.0"
arrow = "^0.16.0"
flask_mongoengine = "^0.9.5"
pytest = "^6.1.1"

[tool.poetry.dev-dependencies]
mongomock = "^3.20.0"
pytest = "^6.1"
pytest-factoryboy = "^2.0.3"
pytest-env = "^0.6.2"
pytest-flask = "^1.0.0"
flask-shell-ipython = "^0.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
